% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatMix-package.R
\docType{package}
\name{SpatMix-package}
\alias{SpatMix}
\alias{SpatMix-package}
\title{SpatMix: Spatial Mixture Models with Matrix Partial EM}
\description{
Fits spatial mixture models, including spatial Gaussian mixtures and mixtures of spatial factor analyzers, to complete or incomplete data. Spatial decay can be represented by monotone I-splines or a normalized sigmoid. Missing entries are handled using a built-in partial expectation-maximization procedure for matrix-variate data. The spatial covariance and spatial factor analyzer models are described in Lu and colleagues (2026a) "Spatial Covariance Constraints for Gaussian Mixture Models" \doi{10.48550/arXiv.2601.07979} and Lu and colleagues (2026b) "Mixtures of spatial factor analyzers for tensor-variate data" \doi{10.48550/arXiv.2607.07887}.
}
\author{
\strong{Maintainer}: Hanzhang Lu \email{hanzhang.lu@ubc.ca}

Authors:
\itemize{
  \item Jeffrey L. Andrews [thesis advisor]
}

}
\keyword{internal}
